Look I give up.
There's just no way to summarize all the changes I'm making into commits
like this. I'll be working on one thing and it'll lead me down a rabbit
hole to fix another, and another, and eventually I've made 7 different
fixes and 13 different changes and they all depend on one another in a
way that's impossible to disentangle. So from now on I'm just going to
lurch forward in huge all-changing commits that make everything a little
better, instead of pretending that I can just fix one thing at a time.