models for Post and Vote, moved code into app
Apparently you can't have models outside of an app. But my code is too
specialized to be "plug-and-play", which is how apps are supposed to
work, right?

In any case, I'll have to have at least one app. I named it "main_app"
because there was already a folder called "hypnohub_post_finder". I may
be able to split everything up into multiple apps at a later date, in
which case I should rename this app to more closely represent what it
does.