Added a 'design goals' file as a todo list
I don't know why I didn't think of this earlier! This is going to help a
lot with staying organized and staying on-task. I can even put each todo
in priority order so that I can have an alpha build sooner.

For those reading, none of the above is actually going to happen. I'm
not that organized. But I can dream.