Met permalink design goals
The Post model now has a 'permalink' property, but the way it works is
kind of an ugly hack. Will have to fix later.

post_view.html has been cleaned up a lot and now every page has an
auto-copy-able permalink box.

Updated design goals to reflect all of this.